requirements.txt
uvicorn==0.32.1
streamlit==1.29.0
requests==2.32.3
fastapi==0.115.5
pydantic==2.9.2
pandas==2.2.3
typing-extensions==4.12.2
python-dotenv==1.0.1
langchain-experimental==0.3.3
langchain-openai==0.2.12
openai==1.57.2
matplotlib==3.9.2


